Quiz App
In this project 5 screen made below is their name and fuctionality
Splash screeen:-
This screen is the landing screen of the application and displays the application logo and only display for 1.5 seconds The SplashScreen transition to the SetupScreen on successful completion of 1.5 seconds. Setup Screen:-
This screen is launched automatically by the SplashScreen Pressing the Start Button on this screen start the timer and transition to the QuestionsListScreen. Pressing Back Button on this screen not launch the SplashScreen but close the application. QuestionListScreen:-
On clicking on a question item in recycler view open questionDetailScreen On pressing back screen close the application
Also have submit button on clicking show the dialog to confir submition On top is text view to show time left Foreground services is used for timer
QuestionDetailScreen:-
Screen showsthe details of question tapped in questionListScreen screen have next Button to goe for next question and previous button to go in previous question and also bookmark to add or remove bookmark save and remove button save button save the answer and shows choosed option and clear button clear the answer on save of answer Question status changes from not answered to answered and on clear it chanes from answered to not answered
and submit button shows dialog to confirm the submission of test
Summary screen:-
shows the time used in exam
shows the score of exam
restart button start the exam from setup screen
exit button exit the application

application Maintains state on orientaion change

Checked used Acessbility scanner to check ex:- color contrast, touch size, item labels

Used Data Binding to views to bind data

Navigation Test are written to check App WorkFlow is corrrect or not and also to check recyclerview Scrolling